commit segunda 2

+++ site/src/api/admin/produtoAPI.js
@@ -4,21 +4,21 @@ const api = axios.create({
 })
 
 export async function buscaPorNome(nome){
-    const resposta = await api.get(`/produto/busca?nome=%${nome}%`);
+    const resposta = await api.get(`/admin/produto/busca?nome=%${nome}%`);
     return resposta.data; 
 }
 
 export async function mostrarTodosProdutos(){
-    const resposta = await api.get('/produto/');
+    const resposta = await api.get('/admin/produto/');
     return resposta.data; 
 }
 
 export async function removerProduto(id){
-    const resposta = await api.delete(`/produto/${id}`);
+    const resposta = await api.delete(`/admin/produto/${id}`);
     return resposta.status;
 }
 
 export async function BuscaporId(id){
-    const resposta = await api.get(`/produto/${id}`);
+    const resposta = await api.get(`/admin/produto/${id}`);
     return resposta.data
 }
